requirements.txt
dgl==1.0.0
easydict==1.10
ema_pytorch==0.2.1
json5==0.9.11
networkx==3.0
numpy==1.24.2
ogb==1.3.5
pandas==1.5.3
PyYAML==6.0
scikit_learn==1.2.1
scipy==1.10.1
torch==1.13.1
torch_geometric==2.2.0
torchmetrics==0.11.1
tqdm==4.64.1
wandb==0.13.10
